+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## Unreleased-->
+## Unreleased
+
+### Added
+
+- logs endpoints
+
 ## [0.22.0] - 2026-03-31
 
 ### Added