+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to
 
 <!--## Unreleased-->
 
+## [0.24.0] - 2026-04-16
+
+### Added 
+
+- `TrackingRecord` and `TrackingCaaRecord` in `ProxyStatus`
+
 ## [0.23.0] - 2026-04-12
 
 ### Added
@@ -483,6 +489,7 @@ Disabled `reqwest`'s default features and enabled `rustls-tls`.
 
 Initial release.
 
+[0.24.0]: https://crates.io/crates/resend-rs/0.24.0
 [0.23.0]: https://crates.io/crates/resend-rs/0.23.0
 [0.22.0]: https://crates.io/crates/resend-rs/0.22.0
 [0.21.1]: https://crates.io/crates/resend-rs/0.21.1